% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createtableoptions.R
\name{add_hash_partitions}
\alias{add_hash_partitions}
\title{add_hash_partitions}
\usage{
add_hash_partitions(opts, columns, buckets, seed = 0)
}
\arguments{
\item{opts}{additional options}

\item{columns}{columns for partitioning}

\item{buckets}{buckets}

\item{seed}{seed}
}
\description{
add hash partitions to kudu table
}
