% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kudu.R
\name{create_kudu_table}
\alias{create_kudu_table}
\title{create_kudu_table}
\usage{
create_kudu_table(sc, table_name, schema, keys, options)
}
\arguments{
\item{sc}{spark connection}

\item{table_name}{table name}

\item{schema}{database in hive}

\item{keys}{key columns}

\item{options}{additional options}
}
\description{
creates a kudu table
}
