% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kudu.R
\name{delete_kudu_table}
\alias{delete_kudu_table}
\title{delete_kudu_table}
\usage{
delete_kudu_table(sc, kudu_table)
}
\arguments{
\item{sc}{spark connection}

\item{kudu_table}{table name}
}
\description{
deletes a kudu table
}
