% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kudu.R
\name{get_impala_ddl}
\alias{get_impala_ddl}
\title{get_impala_ddl}
\usage{
get_impala_ddl(sc, kudu_table)
}
\arguments{
\item{sc}{spark connection}

\item{kudu_table}{kudu table name}
}
\description{
builds impala sql for table creation
}
\examples{
#'  \dontrun{
table_dll <- get_impala_ddl(sc, "table_name")
}

}
