% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kudu.R
\name{get_kudu_table}
\alias{get_kudu_table}
\title{get_kudu_table}
\usage{
get_kudu_table(sc, kudu_table)
}
\arguments{
\item{sc}{spark connection}

\item{kudu_table}{kudu table name}
}
\description{
gets a kudu table
}
