% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kudu.R
\name{kudu_context}
\alias{kudu_context}
\title{kudu_context}
\usage{
kudu_context(sc, kudu_master)
}
\arguments{
\item{sc}{spark connection}

\item{kudu_master}{kudu master}
}
\description{
creates kc kudu context
}
