% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kudu.R
\name{kudu_delete_rows}
\alias{kudu_delete_rows}
\title{kudu_delete_rows}
\usage{
kudu_delete_rows(sc, df, kudu_table)
}
\arguments{
\item{sc}{spark connection}

\item{df}{spark dataframe}

\item{kudu_table}{table name}
}
\description{
deletes rows into a kudu table
}
