% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kudu.R
\name{kudu_insert_rows}
\alias{kudu_insert_rows}
\title{kudu_insert_rows}
\usage{
kudu_insert_rows(sc, df, kudu_table)
}
\arguments{
\item{sc}{spark connection}

\item{df}{spark dataframe}

\item{kudu_table}{table name}
}
\description{
inserts rows into a kudu table
}
