% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kudu.R
\name{kudu_upsert_rows}
\alias{kudu_upsert_rows}
\title{kudu_upsert_rows}
\usage{
kudu_upsert_rows(sc, df, kudu_table)
}
\arguments{
\item{sc}{spark connection}

\item{df}{spark dataframe}

\item{kudu_table}{table name}
}
\description{
upserts rows into a kudu table
}
