% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createtableoptions.R
\name{set_num_replicas}
\alias{set_num_replicas}
\title{set_num_replicas}
\usage{
set_num_replicas(opts, num_replicas)
}
\arguments{
\item{opts}{options}

\item{num_replicas}{integer containing the number of replicas}
}
\description{
set number of replicas
}
