% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createtableoptions.R
\name{set_range_partition_columns}
\alias{set_range_partition_columns}
\title{set_range_partition_columns}
\usage{
set_range_partition_columns(opts, columns)
}
\arguments{
\item{opts}{additional options}

\item{columns}{columns}
}
\description{
set range partition columns for kudu table
}
